working:
	make test
	make pic
	make x86_client
	make android_client
broken:
	make simu

Conflicts:
	CommandIO.c
	Makefile
	WiflyControl.cpp
	WiflyControl.h
	WiflyControlCmd.h
	trace.h
	unittest.h

+++ unittest.h
@@ -38,7 +38,7 @@
 	if(RUN) { \
 		int _errors= FUNC(); \
 		if(_errors != 0) { \
-			printf(#FUNC"() run with %d errors\n", _errors); \
+			printf(#FUNC"() run with %d errors\n", (int)_errors); \
 			numErrors+= _errors; \
 		} \
 		numTests++; \
@@ -52,7 +52,7 @@
 
 #define UnitTestMainBegin(X) size_t numErrors = 0; size_t numSkipped = 0; size_t numTests = 0;
 #define UnitTestMainEnd(X) { \
-	printf("%25s run %2d Tests (%2d skipped | %2d errors)\n", __FILE__, numTests, numSkipped, numErrors); \
+	printf("%25s run %2u Tests (%2u skipped | %2u errors)\n", __FILE__, numTests, numSkipped, numErrors); \
 	return numErrors; \
 }
